+++ e3nn/o3/__init__.py
@@ -6,3 +6,5 @@ from .rotation import *  # noqa
 from .s2grid import *  # noqa
 from .tensor_product import *  # noqa
 from .wigner import *  # noqa
+from .linear import *  # noqa
+from .norm import *  # noqa